+++ src/main/gameObjects/Trail.java
@@ -15,7 +15,7 @@ public class Trail extends GameObject {
     private int height;
     private float life;
 
-    public Trail(int x, int y, ID id, Handler handler, Color color, int width, int height, float life) {
+    public Trail(float x, float y, ID id, Handler handler, Color color, int width, int height, float life) {
         super(x, y, id);
         this.handler = handler;
         this.color = color;
@@ -37,7 +37,7 @@ public class Trail extends GameObject {
         Graphics2D g2d = (Graphics2D) g;
         g2d.setComposite(makeTransparent(alpha));
         g.setColor(color);
-        g.fillRect(x, y, width, height);
+        g.fillRect((int) x, (int) y, width, height);
         g2d.setComposite(makeTransparent(1));
     }
