+++ src/main/utils/Utils.java
@@ -2,7 +2,7 @@ package main.utils;
 
 public class Utils {
 
-    public static int clamp(int var, int min, int max) {
+    public static float clamp(float var, float min, float max) {
         if (var >= max) {
             return var = max;
         } else if (var <= min) {
